Starting to improve install process to ensure client.md5 is generated correctly

+++ applychanges.py
@@ -53,32 +53,33 @@ def merge_tree(root_src_dir, root_dst_dir):
             shutil.copy(src_file, dst_dir)
 
 
-def applychanges(mcp_dir):
+def applychanges(mcp_dir, patch_dir = "patches", backup = True):
     print("Applying Changes...")
 
     mod_src_dir = os.path.join(mcp_dir, "src","minecraft")
     mod_bak_dir = os.path.join(mcp_dir, "src","minecraft-bak")
     org_src_dir = os.path.join(mcp_dir, "src",".minecraft_orig")
     
-    if os.path.exists(mod_src_dir):
+    if backup and os.path.exists(mod_src_dir):
         print("Backing up src/minecraft to src/minecraft-bak")
         shutil.rmtree( mod_bak_dir, True )
         shutil.move( mod_src_dir, mod_bak_dir )
     shutil.copytree( org_src_dir, mod_src_dir, ignore=lambda p,f: [".git"]  )
 
     #apply patches
-    apply_patches( mcp_dir, os.path.join( base_dir, "patches"), mod_src_dir )
+    apply_patches( mcp_dir, os.path.join( base_dir, patch_dir ), mod_src_dir )
     #merge in the new classes
     merge_tree( os.path.join( base_dir, "src" ), mod_src_dir )
     
 if __name__ == '__main__':
     parser = OptionParser()
     parser.add_option('-m', '--mcp-dir', action='store', dest='mcp_dir', help='Path to MCP to use', default=None)
+    parser.add_option('-p', '--patch-dir', action='store', dest='patch_dir', help='Path to base patch dir', default='patches')    
     options, _ = parser.parse_args()
 
     if not options.mcp_dir is None:
-        applychanges(os.path.abspath(options.mcp_dir))
+        applychanges(os.path.abspath(options.mcp_dir), options.patch_dir)
     elif os.path.isfile(os.path.join('..', 'runtime', 'commands.py')):
-        applychanges(os.path.abspath('..'))
+        applychanges(os.path.abspath('..'), options.patch_dir)
     else:
-        applychanges(os.path.abspath(mcp_version))
+        applychanges(os.path.abspath(mcp_version), options.patch_dir)
